Base job execution system implementation

+++ src/service/key_service.cpp
@@ -2,7 +2,7 @@
 
 #include "spdlog/spdlog.h"
 
-#include "herd_common/schema_type.hpp"
+#include "herd/common/model/schema_type.hpp"
 
 #include "service/common_exceptions.hpp"
 #include "utils/file_utils.hpp"
@@ -11,8 +11,10 @@
 namespace fs = std::filesystem;
 
 
-void KeyService::add_key(const UUID& session_uuid, herd::common::SchemaType type, const std::vector<std::byte>& key_data)
+void KeyService::add_key(const herd::common::UUID& session_uuid, herd::common::SchemaType type, const std::vector<std::byte>& key_data)
 {
+	std::unique_lock lock(key_mutex_);
+
 	if(!keys_.contains(session_uuid))
 	{
 		create_directory_for_session(session_uuid);
@@ -32,11 +34,11 @@ void KeyService::add_key(const UUID& session_uuid, herd::common::SchemaType type
 
 	keys_.emplace(
 			session_uuid,
-			KeyEntry{type, key_path}
+			KeyEntry{type, key_path, false}
     );
 }
 
-void KeyService::create_directory_for_session(const UUID& uuid)
+void KeyService::create_directory_for_session(const herd::common::UUID& uuid)
 {
 	const auto session_dir_path = key_storage_dir_ / uuid.as_string();
 	const auto session_dir_status = fs::status(session_dir_path);
@@ -55,8 +57,10 @@ void KeyService::create_directory_for_session(const UUID& uuid)
 	}
 }
 
-void KeyService::remove_key(const UUID& session_uuid, herd::common::SchemaType type)
+void KeyService::remove_key(const herd::common::UUID& session_uuid, herd::common::SchemaType type)
 {
+	std::unique_lock lock(key_mutex_);
+
 	const auto [keys_begin, keys_end] = keys_.equal_range(session_uuid);
 	const auto key_with_type_predicate = [type](const auto& key_entry)
 	{
@@ -65,7 +69,14 @@ void KeyService::remove_key(const UUID& session_uuid, herd::common::SchemaType t
 
 	if(const auto key_iter = std::find_if(keys_begin, keys_end,  key_with_type_predicate); key_iter != keys_end)
 	{
-		fs::remove(key_iter->second.key_path);
+		if(key_iter->second.locks == 0)
+		{
+			fs::remove(key_iter->second.key_path);
+		}
+		else
+		{
+			throw ResourceLockedException("Key locked");
+		}
 	}
 	else
 	{
@@ -73,8 +84,51 @@ void KeyService::remove_key(const UUID& session_uuid, herd::common::SchemaType t
 	}
 }
 
-std::vector<herd::common::SchemaType> KeyService::list_available_keys(const UUID& session_uuid) const
+void KeyService::lock_key(const herd::common::UUID& session_uuid, herd::common::SchemaType type)
 {
+	std::unique_lock lock(key_mutex_);
+
+	const auto [keys_begin, keys_end] = keys_.equal_range(session_uuid);
+	const auto key_with_type_predicate = [type](const auto& key_entry)
+	{
+		return key_entry.second.type == type;
+	};
+
+	if(const auto key_iter = std::find_if(keys_begin, keys_end,  key_with_type_predicate); key_iter != keys_end)
+	{
+		++key_iter->second.locks;
+	}
+	else
+	{
+		throw ObjectNotFoundException("No key of selected type assigned to session");
+	}
+}
+
+
+void KeyService::unlock_key(const herd::common::UUID& session_uuid, herd::common::SchemaType type)
+{
+	std::unique_lock lock(key_mutex_);
+
+	const auto [keys_begin, keys_end] = keys_.equal_range(session_uuid);
+	const auto key_with_type_predicate = [type](const auto& key_entry)
+	{
+		return key_entry.second.type == type;
+	};
+
+	if(const auto key_iter = std::find_if(keys_begin, keys_end,  key_with_type_predicate); key_iter != keys_end)
+	{
+		--key_iter->second.locks;
+	}
+	else
+	{
+		throw ObjectNotFoundException("No key of selected type assigned to session");
+	}
+}
+
+std::vector<herd::common::SchemaType> KeyService::list_available_keys(const herd::common::UUID& session_uuid) const
+{
+	std::shared_lock lock(key_mutex_);
+
 	std::vector<herd::common::SchemaType> types;
 	const auto [keys_begin, keys_end] = keys_.equal_range(session_uuid);
 
@@ -88,8 +142,10 @@ std::vector<herd::common::SchemaType> KeyService::list_available_keys(const UUID
 	return types;
 }
 
-bool KeyService::schema_key_exists_for_session(const UUID& session_uuid, herd::common::SchemaType type) const noexcept
+bool KeyService::schema_key_exists_for_session(const herd::common::UUID& session_uuid, herd::common::SchemaType type) const noexcept
 {
+	std::shared_lock lock(key_mutex_);
+
 	const auto [keys_begin, keys_end] = keys_.equal_range(session_uuid);
 	return std::any_of(keys_begin, keys_end,
 		[type](const auto& entry)
